disable selinium test on arm64
